refactor(test): refactor test cases

+++ server/helpers/test_hooks_helper.js
@@ -8,7 +8,7 @@ import dbConnection from '../db/database';
  */
 export const clearAllTestRecords = (done) => {
   dbConnection.dbConnect('DELETE FROM properties_test;')
-    .then(() => dbConnection.dbConnect('DELETE FROM users_test;'))
+    .then(() => dbConnection.dbConnect('DELETE FROM users_test WHERE email <> $1;', [process.env.ADMIN_EMAIL]))
     .then(() => done())
     .catch(e => done(e));
 };
@@ -19,7 +19,7 @@ export const clearAllTestRecords = (done) => {
  * @param {callback} done
  */
 export const clearTestUsersRecords = (done) => {
-  dbConnection.dbConnect('DELETE FROM users_test;')
+  dbConnection.dbConnect('DELETE FROM users_test WHERE email <> $1;', [process.env.ADMIN_EMAIL])
     .then(() => done())
     .catch(e => done(e));
 };
